+++ src/pyobs/mxd04.py
@@ -435,7 +435,7 @@ class MxD04_L2(object):
         """
         
         if self.syn_time == None:
-            raise ValuError("synoptic time missing, cannot write ODS")
+            raise ValueError("synoptic time missing, cannot write ODS")
             
         # Stop here is no good obs available
         # ----------------------------------