+++ src/pyobs/sampler.py
@@ -8,6 +8,7 @@ import os
 
 import numpy  as np
 import xarray as xr
+import xesmf as xe
 import pandas as pd
 
 from datetime import datetime, timedelta
@@ -15,6 +16,8 @@ from dateutil.parser import parse as isoparser
 from glob import glob
 
 from . import xrctl as xc
+import fsspec
+import dask.distributed
 
 os.environ['HDF5_USE_FILE_LOCKING']='FALSE'
 
@@ -32,8 +35,9 @@ class SamplerError(Exception):
 class STATION(object):
 
     def __init__(self, stations, lons, lats,
-                 dataset, time_range=None, verbose=False,
-                 parallel=True,chunks='auto'):
+                 dataset, time_range=None, times=None, 
+                 verbose=False,
+                 parallel=True,chunks='auto',**kwargs):
         """
         Specifies dataset to be sampled at obs location.
         On input,
@@ -48,7 +52,8 @@ class STATION(object):
                  list,tuple: a list of file names
         time_range: when using a GrADS templates, the time interval
               to generate a list of files.
-
+        times: optional specific times to sample at, 
+              otherwise output at model time resolution
         """
 
         self.verb = verbose
@@ -57,13 +62,18 @@ class STATION(object):
         # -------------------------------------------------
         if isinstance(dataset,xr.Dataset):
             self.ds = dataset # we are good to go...
-
+        elif isinstance(dataset,dask.distributed.Future):
+            self.ds = xr.open_dataset(dataset, engine="zarr",chunks=chunks, backend_kwargs={"consolidated": False})
+        #  if dataset is a parquet referece file path
+        elif dataset[-4:] == 'parq':
+            fs = fsspec.filesystem("reference", fo=dataset, remote_protocol='file', lazy=True)
+            self.ds = xr.open_dataset(fs.get_mapper(), engine="zarr",chunks=chunks, backend_kwargs={"consolidated": False})        
         # If dataset is a list of files...
         # OR GrADS-style ctl
         # OR a glob type of template
         # --------------------------------
         elif isinstance(dataset,(list,tuple,str)):
-            self.ds = xc.open_mfdataset(dataset,time_range=time_range,parallel=parallel,chunks=chunks)
+            self.ds = xc.open_mfdataset(dataset,time_range=time_range,parallel=parallel,chunks=chunks,**kwargs)
 
         else:
             raise SamplerError("Invalid dataset specification.")
@@ -74,9 +84,15 @@ class STATION(object):
         self.lons = xr.DataArray(lons, dims='station',attrs=self.ds.coords['lon'].attrs)
         self.lats = xr.DataArray(lats, dims='station',attrs=self.ds.coords['lat'].attrs)
 
-        # TO DO: when using xESMF for regridding, pre-compute transforms here
-        # -------------------------------------------------------------------
+        if times is not None:
+            self.times = xr.DataArray(times,dims='time')
+        else:
+            self.times = times
 
+        # Use xESMF for regridding, pre-compute transforms here
+        # -------------------------------------------------------------------
+        ds_loc = xr.Dataset({"lon": self.lons, "lat": self.lats})
+        self.regridder = xe.Regridder(self.ds, ds_loc, "bilinear", locstream_out=True)
 
     #--
     def sample(self,Variables=None,method='linear'):
@@ -94,7 +110,11 @@ class STATION(object):
 
         for vn in Variables:
             if self.verb: print('[ ] sampling ',vn)
-            sampled[vn] = self.ds[vn].interp(lon=self.lons,lat=self.lats,method=method)
+            sampled[vn] = self.regridder(self.ds[vn])
+#            sampled[vn] = self.ds[vn].interp(time=self.times,lon=self.lons,lat=self.lats,method=method)
+
+            if self.times is not None:
+                sampled[vn] = sampled[vn].interp(time=self.times,method=method)
 
         return xr.Dataset(sampled).assign_coords({'station': self.stations})
 
@@ -102,7 +122,7 @@ class STATION(object):
 
 class TRAJECTORY(object):
 
-    def __init__(self, times, lons, lats, dataset, parallel=True,chunks='auto',verbose=False):
+    def __init__(self, times, lons, lats, dataset, parallel=True,chunks='auto',verbose=False,**kwargs):
         """
         Specifies dataset to be sampled at obs location.
         On input,
@@ -129,13 +149,16 @@ class TRAJECTORY(object):
         # -------------------------------------------------
         if isinstance(dataset,xr.Dataset):
             self.ds = dataset # we are good to go...
-
+        #  if dataset is a parquet referece file
+        elif dataset[-4:] == 'parq':
+            fs = fsspec.filesystem("reference", fo=dataset, remote_protocol='file', lazy=True)
+            self.ds = xr.open_dataset(fs.get_mapper(), engine="zarr", chunks=chunks, backend_kwargs={"consolidated": False})
         # If dataset is a list of files...
         # OR GrADS-style ctl 
         # OR a glob type of template
         # --------------------------------
         elif isinstance(dataset,(list,tuple,str)):
-            self.ds = xc.open_mfdataset(dataset,time_range=time_range,parallel=parallel,chunks=chunks) # special handles GrADS-style ctl if found
+            self.ds = xc.open_mfdataset(dataset,time_range=time_range,parallel=parallel,chunks=chunks,**kwargs) # special handles GrADS-style ctl if found
 
         else:
             raise SamplerError("Invalid dataset specification.")
@@ -393,17 +416,19 @@ def addVertCoord(aer):
         dz = rhodz / aer['AIRDENS']       # column thickness in m
 
         # add up the thicknesses to get edge level altitudes
-        npts, nlev = dz.shape
-        ze = np.array([dz[:,i:].sum(axis=1) for i in range(nlev)])
+        nlev = dz.sizes['lev']
+        ze = xr.concat([dz.isel(lev=slice(i,None)).sum(dim='lev',keep_attrs=True) for i in range(nlev)],dim='lev')
         # append surface level, altitude = 0
-        ze = np.append(ze,np.zeros([1,npts]),axis=0)
-        # transpose to get npts, nlev again
-        ze = ze.T
+        surface = xr.DataArray(np.zeros((1,) +ze.shape[1:]),dims=ze.dims)
+        ze = xr.concat([ze,surface],dim='lev')
+        # transpose back to npts,nlev again
+        dims = list(dz.dims)
+        ze = ze.transpose(*dims)
         # convert from m to km
         ze = ze*1e-3
 
         # get mid-level altitudes
-        z = (ze[:,:-1] + ze[:,1:])*0.5
+        z = (ze.isel(lev=slice(None,-1)) + ze.isel(lev=slice(1,None)))*0.5
 
         # Attributes
         # ----------
@@ -413,8 +438,8 @@ def addVertCoord(aer):
 
         # Pack results into a DataArray
         # ---------------------------
-        DA = dict(  Z = xr.DataArray(z.astype('float32'),dims=dp.dims,coords=dp.coords,attrs=A['Z']),
-                    DZ = xr.DataArray(dz.astype('float32'),dims=dp.dims,coords=dp.coords,attrs=A['DZ'])
+        DA = dict(  Z = z.assign_attrs(A['Z']),
+                    DZ = dz.assign_attrs(A['DZ'])
                  )
 
         # Add to Dataset